Generic database (UNKNOWN database type decommissioning)

+++ src/com/dci/intellij/dbn/database/common/metadata/def/DBDataTypeMetadata.java
@@ -10,7 +10,7 @@ public interface DBDataTypeMetadata {
 
     String getDataTypeProgram() throws SQLException;
 
-    long getDataLength() throws SQLException;
+    int getDataLength() throws SQLException;
 
     int getDataPrecision() throws SQLException;
